feat(minha-conta): componentizar cards informativos
